Add Cloudflare edge middleware for HTTP response headers on static pages
Move `noIndex` and `securityHeaders` middleware from `src/middleware.ts`
into a new dedicated file `src/middlewares/securityHeaders.ts`. This file
acts as the Cloudflare edge worker middleware for HTTP response headers,
keeping it separate from server-side logic (redirects, logging, JWT auth).

Enable `run_worker_first: true` in the wrangler.jsonc assets binding so
the Cloudflare Worker is invoked for every request — including
prerendered/static pages — allowing the security headers middleware to
apply to all responses, not just SSR routes.
